refactor(attendance): AttendanceHistory sous 300 lignes (H7)
Éclatement complet (pattern G1) : la vue (1078 l, Options API) devient un
orchestrateur <script setup> de 188 l. Logique/état/export CSV extraits dans
useAttendanceHistory (215). UI en sous-composants présentationnels :
AttendanceHistoryFilters (152), Stats (103), Table (103) → Rows (238) +
Pagination (82), AttendanceParticipationModal (204). CSS déplacé verbatim par
section (médias inclus).

Parité : formateurs date/heure LOCAUX (heure AVEC secondes) conservés à
l'identique dans le composable et passés en props aux tableaux/modale — la
duplication préexistante vs @/utils/formatters est volontairement gardée pour
un rendu strictement inchangé. CSS mort .loading-container conservé verbatim.

Tests : composable (services mockés) + montage des sous-composants (15 verts).
